# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Claw Admin API
  description: Unified admin control plane for LLM/Skill/MCP config, SSO, organizations, users, and logs.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1 base path
...
security:
  - BearerAuth: []
paths:
...
  /admin/skills:
    get:
      tags: [admin-skills]
      summary: List skills
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      responses:
        '200':
          description: List of skills
    post:
      tags: [admin-skills]
      summary: Create/register skill
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                skillId: { type: string }
                name: { type: string }
                enabled: { type: boolean }
                sortOrder: { type: integer }
                config: { type: object }
      responses:
        '201':
          description: Created
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/skills/{id}:
    get:
      tags: [admin-skills]
      summary: Get skill
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Skill
    put:
      tags: [admin-skills]
      summary: Update skill
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                sortOrder: { type: integer }
                config: { type: object }
      responses:
        '200':
          description: Updated
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [admin-skills]
      summary: Delete skill config
      parameters:
        - $ref: '#/components/parameters/XOrgId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /client/config/skills:
    get:
      tags: [client]
      summary: Get current tenant skills
      parameters:
        - $ref: '#/components/parameters/XOrgId'
      responses:
        '200':
          description: Skill list with config
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    XOrgId:
      name: X-Org-Id
      in: header
      description: Tenant slug (optional; can come from JWT org_id)
      schema:
        type: string
        example: default
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: UNAUTHORIZED }
                  message: { type: string }
                  details: { type: object }
    Forbidden:
      description: Forbidden (admin role required)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: FORBIDDEN }
                  message: { type: string }
...